Delay first rising clock edge by 1 half-period in VPI Harness
This prevents a race condition between the rising clock edge and
randomization when RANDOMIZE_DELAY=1. Note that the traditional Chisel
randomization delay of #0.002 is based on VCS's default timescale of
1ns/1ps and does not properly work with other timescales. It is much
safer to use a full timescale delay, and thus prudent to delay the first
rising clock edge till the 2nd time step.