Remove Icarus debug messages (#607)
When using Icarus backend there was many debug messages that was not needed.
I changed it to not output anything normally but show debug messages when "SimulatorDebugAnnotation" is set, like the Verilator backend.

By the way I found that some tests in `PeekPokeCompliance` had not properly finished the dut. it was causing the simulator to keep running after the test finished.